fix circulair dependencies
